requirements.txt
streamlit>=1.28.0
openai>=1.0.0
langchain>=0.1.0
langchain-community>=0.0.10
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0
PyPDF2>=3.0.0
python-docx>=0.8.11
tiktoken>=0.5.0
requests>=2.31.0
numpy>=1.24.0
pandas>=2.0.0
python-dotenv>=1.0.0
scikit-learn>=1.3.0